Refactoring; Improved unit tests.
Extracted the data-retrieval code to a new helper class. Also updated
and improved the unit tests.
