+++ chebi-update/src/main/java/org/reactome/release/chebiupdate/ChebiUpdater.java
@@ -1,13 +1,7 @@
 package org.reactome.release.chebiupdate;
 
-import java.io.BufferedWriter;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
@@ -16,7 +10,6 @@ import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
 import org.apache.logging.log4j.LogManager;
@@ -30,8 +23,6 @@ import org.gk.schema.InvalidAttributeException;
 import org.gk.schema.InvalidAttributeValueException;
 import org.reactome.release.common.database.InstanceEditUtils;
 
-import uk.ac.ebi.chebi.webapps.chebiWS.client.ChebiWebServiceClient;
-import uk.ac.ebi.chebi.webapps.chebiWS.model.ChebiWebServiceFault_Exception;
 import uk.ac.ebi.chebi.webapps.chebiWS.model.DataItem;
 import uk.ac.ebi.chebi.webapps.chebiWS.model.Entity;
 
@@ -50,7 +41,7 @@ public class ChebiUpdater
 	private static final Logger failedChebiLookupsLog = LogManager.getLogger("failedChebiLookupsLog");
 	private boolean testMode = true;
 	private MySQLAdaptor adaptor;
-	private ChebiWebServiceClient chebiClient = new ChebiWebServiceClient();
+//	private ChebiWebServiceClient chebiClient = new ChebiWebServiceClient();
 	private StringBuilder formulaUpdateSB = new StringBuilder();
 	private StringBuilder formulaFillSB = new StringBuilder();
 	private StringBuilder duplicatesSB = new StringBuilder();
@@ -134,7 +125,8 @@ public class ChebiUpdater
 
 		// A map: key is the DB_ID of a ReferneceMolecule, value is the
 		// uk.ac.ebi.chebi.webapps.chebiWS.model.Entity from ChEBI.
-		Map<Long, Entity> entityMap = retrieveUpdatesFromChebi(refMolecules, failedEntitiesMap);
+		ChebiDataRetriever dataRetriever = new ChebiDataRetriever(this.useCache);
+		Map<Long, Entity> entityMap = dataRetriever.retrieveUpdatesFromChebi(refMolecules, failedEntitiesMap);
 
 		logger.info("Number of entities we were able to retrieve information about: {}", entityMap.size());
 		logger.info("Number of entities we were NOT able to retrieve information about: {}", failedEntitiesMap.size());
@@ -143,7 +135,7 @@ public class ChebiUpdater
 		for (GKInstance molecule : failedEntitiesMap.keySet())
 		{
 			GKInstance creator = ChebiUpdater.getCreator(molecule);
-			failedChebiLookupsLog.info("{}\t{}\t{}\t{}\t{}", molecule.getDBID(), creator.toString(), molecule.toString(), molecule.toString(), failedEntitiesMap.get(molecule));
+			failedChebiLookupsLog.info("{}\t{}\t{}\t{}", molecule.getDBID(), creator.toString(), molecule.toString(), failedEntitiesMap.get(molecule));
 		}
 		
 		// print headers for log files
@@ -506,182 +498,182 @@ public class ChebiUpdater
 		}
 	}
 
-	/**
-	 * This looks weird, I know. I needed to be able to set the Formulae on an Entity 
-	 * and the Entity class provided by ChEBI does not have a setter for that. It has setters for
-	 * other members, just not all of them. So I added a setter, hence the "accessible" name.
-	 * @author sshorser
-	 *
-	 */
-	private class AccessibleEntity extends Entity
-	{
-		public void setFormulae(List<DataItem> formulae)
-		{
-			this.formulae = formulae;
-		}
-	}
-	
-	/**
-	 * Makes calls to the ChEBI web service to get info for speciefied ChEBI
-	 * identifiers.
-	 * 
-	 * @param updator
-	 * @param refMolecules - a list of ReferenceMolecules. The Identifier of each of these molecules will be sent to ChEBI to get up-to-date information for that Identifier.
-	 * @param failedEntitiesList - A list of ReferenceMolecules for which no information was returned by ChEBI. Will be updated by this method.
-	 * @return A ReferenceMolecule DB_ID-to-ChEBI Entity map.
-	 * @throws IOException 
-	 */
-	private Map<Long, Entity> retrieveUpdatesFromChebi(Collection<GKInstance> refMolecules, Map<GKInstance, String> failedEntitiesList) throws IOException
-	{
-		Map<Long, Entity> entityMap = Collections.synchronizedMap(new HashMap<Long, Entity>());
-		final Map<String,List<String>> chebiCache = loadCacheFromFile();
-		
-		// BufferedWriter is supposed to be thread-safe.
-		try(FileWriter fileWriter = new FileWriter("chebi-cache", true);
-			BufferedWriter bw = new BufferedWriter(fileWriter);)
-		{
-			AtomicInteger counter = new AtomicInteger(0);
-			// The web service calls are a bit slow to respond, so do them in parallel.
-			refMolecules.parallelStream().forEach(molecule ->
-			{
-				String identifier = null;
-				try
-				{
-					identifier = (String) molecule.getAttributeValue("identifier");
-					if (identifier != null && !identifier.trim().equals(""))
-					{
-						Entity entity;
-						// only query web service if the data is not in the chebi-cache - NOTE: chebiCache will always be empty if this.useCache == false
-						if (!chebiCache.containsKey("CHEBI:"+identifier))
-						{
-							if (this.useCache && chebiCache.size() > 0)
-							{
-								logger.trace("Cache miss for CHEBI:{}", identifier);
-							}
-							entity = this.chebiClient.getCompleteEntity(identifier);
-							if (entity != null && this.useCache)
-							{
-								bw.write("CHEBI:"+identifier+"\t"+entity.getChebiId()+"\t"+entity.getChebiAsciiName()+"\t"+ (entity.getFormulae().size() > 0 ? entity.getFormulae().get(0).getData() : "") + "\t" + LocalDateTime.now().toString() + "\n");
-								bw.flush();
-							}
-							else
-							{
-								if (entity == null)
-								{
-									failedEntitiesList.put(molecule, "ChEBI WebService response was NULL.");
-								}
-							}
-						}
-						else // ...Load data from the cache.
-						{
-							entity = extractChEBIEntityFromCache(chebiCache, identifier);
-						}
-						// Add entity to map (if it's non-null)
-						if (entity != null)
-						{
-							entityMap.put(molecule.getDBID(), entity);
-						}
-					}
-					else
-					{
-						logger.error("ERROR: Instance \"{}\" has an empty/null identifier. This should not be allowed.", molecule.toString());
-						failedEntitiesList.put(molecule, molecule.toString() + " has an empty/NULL identifier.");
-					}
-					int i = counter.getAndIncrement();
-					if (i % 250 == 0)
-					{
-						logger.debug("{} ChEBI identifiers checked", i);
-					}
-				}
-				catch (ChebiWebServiceFault_Exception e)
-				{
-					// "invalid ChEBI identifier" probably shouldn't break execution but should be logged for further investigation.
-					if (e.getMessage().contains("invalid ChEBI identifier"))
-					{
-						logger.error("ERROR: ChEBI Identifier \"{}\" is not formatted correctly.", identifier);
-						failedEntitiesList.put(molecule, "ChEBI Identifier \""+identifier+"\" is not formatted correctly.");
-					}
-					// Log this identifier, but don't fail.
-					else if (e.getMessage().contains("the entity in question is deleted, obsolete, or not yet released"))
-					{
-						logger.error("ERROR: ChEBI Identifier \"{}\" is deleted, obsolete, or not yet released.", identifier);
-						failedEntitiesList.put(molecule, "ChEBI Identifier \""+identifier+"\" is deleted, obsolete, or not yet released.");
-					}
-					else
-					{
-						// Other Webservice errors should probably break execution - if one fails, they will all probably fail.
-						// This is *not* a general principle, but is based on my experience with the ChEBI webservice specifically -
-						// it's a pretty stable service so it's unlikely that if one service call fails, the others will succeed.
-						logger.error("WebService error occurred! Message is: {}", e.getMessage());
-						e.printStackTrace();
-						throw new RuntimeException(e);
-					}
-				}
-				catch (InvalidAttributeException e)
-				{
-					logger.error("InvalidAttributeException caught while trying to get the \"identifier\" attribute on " + molecule.toString());
-					// stack trace should be printed, but I don't think this should break execution, though the only way I can think
-					// of this happening is if the data model changes - otherwise, this exception would probably never be caught.
-					e.printStackTrace();
-				}
-				catch (Exception e)
-				{
-					// general exceptions - print stack trace but keep going.
-					e.printStackTrace();
-				}
-			});
-		}
-		return entityMap;
-	}
-
-	/**
-	 * Returns the data from the cache file (if it exists). If it doesn't exit, then an empty map will be returned.
-	 * @return A mapping of ChEBI ID to a list with items in this order: ChEBI (might be different if ChEBI has changed identifiers), ChEBI Name, Formula.
-	 * @throws IOException
-	 */
-	private Map<String,List<String>> loadCacheFromFile() throws IOException
-	{
-		Map<String,List<String>> chebiCache = Collections.synchronizedMap(new HashMap<String, List<String>>());
-		if (this.useCache)
-		{
-			logger.info("useCache is TRUE - chebi-cache file will be read, and populated. Identifiers not in the cache will be queried from ChEBI.");
-			// if the cache exists, load it.
-			if (Files.exists(Paths.get("chebi-cache")))
-			{
-				Files.readAllLines(Paths.get("chebi-cache")).parallelStream().forEach( line -> {
-					String[] parts = line.split("\t");
-					String oldChebiID = parts[0];
-					String newChebiID = parts[1];
-					String name = parts[2];
-					String formula = parts.length > 2 ? parts[3] : "";
-					chebiCache.put(oldChebiID, Arrays.asList(newChebiID, name, formula));
-				});
-			}
-			logger.debug("{} entries in the chebi-cache", chebiCache.size());
-		}
-		else
-		{
-			logger.info("useCache is FALSE - chebi-cache will NOT be read. ChEBI will be queried for ALL identifiers.");
-		}
-		return chebiCache;
-	}
-
-	/**
-	 * Extract a ChEBI entity from the cache, based on the identifier.
-	 * @param chebiCache - The cache from a file.
-	 * @param identifier - The chebi Identifier.
-	 * @return A ChEBI Entity that will be built from the data in the cache. It will only have: an Identifier, ChEBI Name, Formula.
-	 */
-	private Entity extractChEBIEntityFromCache(Map<String, List<String>> chebiCache, String identifier)
-	{
-		Entity entity;
-		// Use the AccessibleEntity to set the formula.
-		entity = new AccessibleEntity();
-		entity.setChebiId(chebiCache.get("CHEBI:"+identifier).get(0));
-		entity.setChebiAsciiName(chebiCache.get("CHEBI:"+identifier).get(1) );
-		DataItem formula = new DataItem();
-		formula.setData(chebiCache.get("CHEBI:"+identifier).get(2));
-		((AccessibleEntity)entity).setFormulae(Arrays.asList(formula));
-		return entity;
-	}
+//	/**
+//	 * This looks weird, I know. I needed to be able to set the Formulae on an Entity 
+//	 * and the Entity class provided by ChEBI does not have a setter for that. It has setters for
+//	 * other members, just not all of them. So I added a setter, hence the "accessible" name.
+//	 * @author sshorser
+//	 *
+//	 */
+//	private class AccessibleEntity extends Entity
+//	{
+//		public void setFormulae(List<DataItem> formulae)
+//		{
+//			this.formulae = formulae;
+//		}
+//	}
+//	
+//	/**
+//	 * Makes calls to the ChEBI web service to get info for speciefied ChEBI
+//	 * identifiers.
+//	 * 
+//	 * @param updator
+//	 * @param refMolecules - a list of ReferenceMolecules. The Identifier of each of these molecules will be sent to ChEBI to get up-to-date information for that Identifier.
+//	 * @param failedEntitiesList - A list of ReferenceMolecules for which no information was returned by ChEBI. Will be updated by this method.
+//	 * @return A ReferenceMolecule DB_ID-to-ChEBI Entity map.
+//	 * @throws IOException 
+//	 */
+//	private Map<Long, Entity> retrieveUpdatesFromChebi(Collection<GKInstance> refMolecules, Map<GKInstance, String> failedEntitiesList) throws IOException
+//	{
+//		Map<Long, Entity> entityMap = Collections.synchronizedMap(new HashMap<Long, Entity>());
+//		final Map<String,List<String>> chebiCache = loadCacheFromFile();
+//		
+//		// BufferedWriter is supposed to be thread-safe.
+//		try(FileWriter fileWriter = new FileWriter("chebi-cache", true);
+//			BufferedWriter bw = new BufferedWriter(fileWriter);)
+//		{
+//			AtomicInteger counter = new AtomicInteger(0);
+//			// The web service calls are a bit slow to respond, so do them in parallel.
+//			refMolecules.parallelStream().forEach(molecule ->
+//			{
+//				String identifier = null;
+//				try
+//				{
+//					identifier = (String) molecule.getAttributeValue("identifier");
+//					if (identifier != null && !identifier.trim().equals(""))
+//					{
+//						Entity entity;
+//						// only query web service if the data is not in the chebi-cache - NOTE: chebiCache will always be empty if this.useCache == false
+//						if (!chebiCache.containsKey("CHEBI:"+identifier))
+//						{
+//							if (this.useCache && chebiCache.size() > 0)
+//							{
+//								logger.trace("Cache miss for CHEBI:{}", identifier);
+//							}
+//							entity = this.chebiClient.getCompleteEntity(identifier);
+//							if (entity != null && this.useCache)
+//							{
+//								bw.write("CHEBI:"+identifier+"\t"+entity.getChebiId()+"\t"+entity.getChebiAsciiName()+"\t"+ (entity.getFormulae().size() > 0 ? entity.getFormulae().get(0).getData() : "") + "\t" + LocalDateTime.now().toString() + "\n");
+//								bw.flush();
+//							}
+//							else
+//							{
+//								if (entity == null)
+//								{
+//									failedEntitiesList.put(molecule, "ChEBI WebService response was NULL.");
+//								}
+//							}
+//						}
+//						else // ...Load data from the cache.
+//						{
+//							entity = extractChEBIEntityFromCache(chebiCache, identifier);
+//						}
+//						// Add entity to map (if it's non-null)
+//						if (entity != null)
+//						{
+//							entityMap.put(molecule.getDBID(), entity);
+//						}
+//					}
+//					else
+//					{
+//						logger.error("ERROR: Instance \"{}\" has an empty/null identifier. This should not be allowed.", molecule.toString());
+//						failedEntitiesList.put(molecule, molecule.toString() + " has an empty/NULL identifier.");
+//					}
+//					int i = counter.getAndIncrement();
+//					if (i % 250 == 0)
+//					{
+//						logger.debug("{} ChEBI identifiers checked", i);
+//					}
+//				}
+//				catch (ChebiWebServiceFault_Exception e)
+//				{
+//					// "invalid ChEBI identifier" probably shouldn't break execution but should be logged for further investigation.
+//					if (e.getMessage().contains("invalid ChEBI identifier"))
+//					{
+//						logger.error("ERROR: ChEBI Identifier \"{}\" is not formatted correctly.", identifier);
+//						failedEntitiesList.put(molecule, "ChEBI Identifier \""+identifier+"\" is not formatted correctly.");
+//					}
+//					// Log this identifier, but don't fail.
+//					else if (e.getMessage().contains("the entity in question is deleted, obsolete, or not yet released"))
+//					{
+//						logger.error("ERROR: ChEBI Identifier \"{}\" is deleted, obsolete, or not yet released.", identifier);
+//						failedEntitiesList.put(molecule, "ChEBI Identifier \""+identifier+"\" is deleted, obsolete, or not yet released.");
+//					}
+//					else
+//					{
+//						// Other Webservice errors should probably break execution - if one fails, they will all probably fail.
+//						// This is *not* a general principle, but is based on my experience with the ChEBI webservice specifically -
+//						// it's a pretty stable service so it's unlikely that if one service call fails, the others will succeed.
+//						logger.error("WebService error occurred! Message is: {}", e.getMessage());
+//						e.printStackTrace();
+//						throw new RuntimeException(e);
+//					}
+//				}
+//				catch (InvalidAttributeException e)
+//				{
+//					logger.error("InvalidAttributeException caught while trying to get the \"identifier\" attribute on " + molecule.toString());
+//					// stack trace should be printed, but I don't think this should break execution, though the only way I can think
+//					// of this happening is if the data model changes - otherwise, this exception would probably never be caught.
+//					e.printStackTrace();
+//				}
+//				catch (Exception e)
+//				{
+//					// general exceptions - print stack trace but keep going.
+//					e.printStackTrace();
+//				}
+//			});
+//		}
+//		return entityMap;
+//	}
+//
+//	/**
+//	 * Returns the data from the cache file (if it exists). If it doesn't exit, then an empty map will be returned.
+//	 * @return A mapping of ChEBI ID to a list with items in this order: ChEBI (might be different if ChEBI has changed identifiers), ChEBI Name, Formula.
+//	 * @throws IOException
+//	 */
+//	private Map<String,List<String>> loadCacheFromFile() throws IOException
+//	{
+//		Map<String,List<String>> chebiCache = Collections.synchronizedMap(new HashMap<String, List<String>>());
+//		if (this.useCache)
+//		{
+//			logger.info("useCache is TRUE - chebi-cache file will be read, and populated. Identifiers not in the cache will be queried from ChEBI.");
+//			// if the cache exists, load it.
+//			if (Files.exists(Paths.get("chebi-cache")))
+//			{
+//				Files.readAllLines(Paths.get("chebi-cache")).parallelStream().forEach( line -> {
+//					String[] parts = line.split("\t");
+//					String oldChebiID = parts[0];
+//					String newChebiID = parts[1];
+//					String name = parts[2];
+//					String formula = parts.length > 2 ? parts[3] : "";
+//					chebiCache.put(oldChebiID, Arrays.asList(newChebiID, name, formula));
+//				});
+//			}
+//			logger.debug("{} entries in the chebi-cache", chebiCache.size());
+//		}
+//		else
+//		{
+//			logger.info("useCache is FALSE - chebi-cache will NOT be read. ChEBI will be queried for ALL identifiers.");
+//		}
+//		return chebiCache;
+//	}
+//
+//	/**
+//	 * Extract a ChEBI entity from the cache, based on the identifier.
+//	 * @param chebiCache - The cache from a file.
+//	 * @param identifier - The chebi Identifier.
+//	 * @return A ChEBI Entity that will be built from the data in the cache. It will only have: an Identifier, ChEBI Name, Formula.
+//	 */
+//	private Entity extractChEBIEntityFromCache(Map<String, List<String>> chebiCache, String identifier)
+//	{
+//		Entity entity;
+//		// Use the AccessibleEntity to set the formula.
+//		entity = new AccessibleEntity();
+//		entity.setChebiId(chebiCache.get("CHEBI:"+identifier).get(0));
+//		entity.setChebiAsciiName(chebiCache.get("CHEBI:"+identifier).get(1) );
+//		DataItem formula = new DataItem();
+//		formula.setData(chebiCache.get("CHEBI:"+identifier).get(2));
+//		((AccessibleEntity)entity).setFormulae(Arrays.asList(formula));
+//		return entity;
+//	}
 }
